code format; rename the scripts; add json parser
