+++ .gitignore
@@ -48,5 +48,6 @@ AGENTS.md
 
 test/BB_code_benchmark
 deprecated_code
+Review_report.md
 
 # sinter_example.py
