+++ Makefile
@@ -1,4 +1,4 @@
-.PHONY: typecheck mypy test check
+.PHONY: typecheck mypy test check format style
 
 typecheck:
 	mypy --config-file mypy.ini .
@@ -8,4 +8,10 @@ mypy: typecheck
 test:
 	python -m unittest discover -s test -p "test_*.py" -v
 
-check: typecheck test
+style:
+	black --check --diff .
+
+format:
+	black .
+
+check: style typecheck test
