+++ config_examples/bb_small_test.json
@@ -1,0 +1,21 @@
+{
+  "description": "Small BB code test configuration for quick verification",
+  "a_poly": [[3, 0], [0, 1], [0, 2]],
+  "b_poly": [[0, 3], [1, 0], [2, 0]],
+  "l": 6,
+  "m": 6,
+  "p_range": {
+    "min": 0.001,
+    "max": 0.007,
+    "num_points": 3
+  },
+  "rounds_list": [6, 8],
+  "max_shots": 1000,
+  "max_errors": 10,
+  "seed": 42,
+  "bp_iters": 10,
+  "osd_order": 0,
+  "num_workers": 2,
+  "resume_every": 100,
+  "resume_csv": "Data/bb_6_6_test_resume.csv"
+}
